+++ CHANGELOG.md
@@ -62,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#505](https://github.com/alleslabs/celatone-frontend/pull/505) Adjust attach funds form label and icon styling for schema section
 - [#501](https://github.com/alleslabs/celatone-frontend/pull/501) Add more JSON Schema state, e.g. empty object state, boolean field
 - [#502](https://github.com/alleslabs/celatone-frontend/pull/502) Display queried time and add json/schema output switch
 - [#500](https://github.com/alleslabs/celatone-frontend/pull/500) Disable estimated fee when input is invalid